Remove establishes_ifc, and print whether we're a bfc in tests.
establishes_ifc is a bit tricky, because it's not exclusive with
establishes_bfc, and also because we definitely have inlines and blocks as the
same child of a block, for example.

Given we don't yet have a use for it per se, and that I think it's not a very
useful concept per se, let's just remove it.